added pagination

+++ src/main/java/org/dnu/filestorage/controller/generic/GenericController.java
@@ -42,7 +42,7 @@ public abstract class GenericController<S extends GenericService<T>, T extends I
 
     @RequestMapping(value = "/count")
     @ResponseBody
-    public Integer getCount() {
+    public Long getCount() {
         return this.service.getCount();
     }
